Refactor roles the "ansible way"
"role" is a misnomer, it's closer to a task packaged with resources that task
needs. More "action" than "role". We need one role for each discrete action we
want to be able to take if we want to be able to perform that action in differnt
contexts, from different places.

Even then it's painfully limiting because one role can't invoke another role
midway through. You have to include the role's tasks, and the tasks aren't
searched for in depended-on roles, so you have to use full paths.